Remove Header/Footer properties from CarouselView
