+++ Xamarin.Forms.Core/Items/ItemsView.cs
@@ -93,42 +93,6 @@ namespace Xamarin.Forms
 			set => SetValue(RemainingItemsThresholdProperty, value);
 		}
 
-		public static readonly BindableProperty HeaderProperty =
-			BindableProperty.Create(nameof(Header), typeof(object), typeof(ItemsView), null);
-
-		public object Header
-		{
-			get => GetValue(HeaderProperty);
-			set => SetValue(HeaderProperty, value);
-		}
-
-		public static readonly BindableProperty HeaderTemplateProperty =
-			BindableProperty.Create(nameof(HeaderTemplate), typeof(DataTemplate), typeof(ItemsView), null);
-
-		public DataTemplate HeaderTemplate
-		{
-			get => (DataTemplate)GetValue(HeaderTemplateProperty);
-			set => SetValue(HeaderTemplateProperty, value);
-		}
-
-		public static readonly BindableProperty FooterProperty =
-			BindableProperty.Create(nameof(Footer), typeof(object), typeof(ItemsView), null);
-
-		public object Footer
-		{
-			get => GetValue(FooterProperty);
-			set => SetValue(FooterProperty, value);
-		}
-
-		public static readonly BindableProperty FooterTemplateProperty =
-			BindableProperty.Create(nameof(FooterTemplate), typeof(DataTemplate), typeof(ItemsView), null);
-
-		public DataTemplate FooterTemplate
-		{
-			get => (DataTemplate)GetValue(FooterTemplateProperty);
-			set => SetValue(FooterTemplateProperty, value);
-		}
-
 		public void AddLogicalChild(Element element)
 		{
 			if(element == null)
@@ -172,15 +136,6 @@ namespace Xamarin.Forms
 			set => SetValue(ItemsLayoutProperty, value);
 		}
 
-		public static readonly BindableProperty ItemTemplateProperty =
-			BindableProperty.Create(nameof(ItemTemplate), typeof(DataTemplate), typeof(ItemsView));
-
-		public DataTemplate ItemTemplate
-		{
-			get => (DataTemplate)GetValue(ItemTemplateProperty);
-			set => SetValue(ItemTemplateProperty, value);
-		}
-
 		public static readonly BindableProperty ItemSizingStrategyProperty =
 			BindableProperty.Create(nameof(ItemSizingStrategy), typeof(ItemSizingStrategy), typeof(ItemsView));
 
@@ -188,6 +143,16 @@ namespace Xamarin.Forms
 		{
 			get => (ItemSizingStrategy)GetValue(ItemSizingStrategyProperty);
 			set => SetValue(ItemSizingStrategyProperty, value);
+		}
+
+
+		public static readonly BindableProperty ItemTemplateProperty =
+			BindableProperty.Create(nameof(ItemTemplate), typeof(DataTemplate), typeof(ItemsView));
+
+		public DataTemplate ItemTemplate
+		{
+			get => (DataTemplate)GetValue(ItemTemplateProperty);
+			set => SetValue(ItemTemplateProperty, value);
 		}
 
 		public static readonly BindableProperty ItemsUpdatingScrollModeProperty =
