+++ Xamarin.Forms.Platform.Android/CollectionView/ItemsViewAdapter.cs
@@ -33,9 +33,6 @@ namespace Xamarin.Forms.Platform.Android
 			_createItemContentView = createItemContentView;
 			ItemsSource = CreateItemsSource();
 
-			UpdateHasHeader();
-			UpdateHasFooter();
-
 			if (_createItemContentView == null)
 			{
 				_createItemContentView = (view, context) => new ItemContentView(context);
@@ -53,17 +50,9 @@ namespace Xamarin.Forms.Platform.Android
 			{
 				UpdateItemsSource();
 			}
-			else if (property.Is(Xamarin.Forms.ItemsView.HeaderProperty))
-			{
-				UpdateHasHeader();
-			}
 			else if (property.Is(Xamarin.Forms.ItemsView.ItemTemplateProperty))
 			{
 				UpdateUsingItemTemplate();
-			}
-			else if (property.Is(Xamarin.Forms.ItemsView.FooterProperty))
-			{
-				UpdateHasFooter();
 			}
 		}
 
@@ -79,26 +68,6 @@ namespace Xamarin.Forms.Platform.Android
 
 		public override void OnBindViewHolder(RecyclerView.ViewHolder holder, int position)
 		{
-			if (IsHeader(position))
-			{
-				if (holder is TemplatedItemViewHolder templatedItemViewHolder)
-				{
-					BindTemplatedItemViewHolder(templatedItemViewHolder, ItemsView.Header);
-				}
-
-				return;
-			}
-
-			if (IsFooter(position))
-			{
-				if (holder is TemplatedItemViewHolder templatedItemViewHolder)
-				{
-					BindTemplatedItemViewHolder(templatedItemViewHolder, ItemsView.Footer);
-				}
-
-				return;
-			}
-
 			switch (holder)
 			{
 				case TextViewHolder textViewHolder:
@@ -113,16 +82,6 @@ namespace Xamarin.Forms.Platform.Android
 		public override RecyclerView.ViewHolder OnCreateViewHolder(ViewGroup parent, int viewType)
 		{
 			var context = parent.Context;
-
-			if (viewType == ItemViewType.Header)
-			{
-				return CreateHeaderFooterViewHolder(ItemsView.Header, ItemsView.HeaderTemplate, context);
-			}
-
-			if (viewType == ItemViewType.Footer)
-			{
-				return CreateHeaderFooterViewHolder(ItemsView.Footer, ItemsView.FooterTemplate, context);
-			}
 
 			if (viewType == ItemViewType.TextItem)
 			{
@@ -139,16 +98,6 @@ namespace Xamarin.Forms.Platform.Android
 
 		public override int GetItemViewType(int position)
 		{
-			if (IsHeader(position))
-			{
-				return ItemViewType.Header;
-			}
-
-			if (IsFooter(position))
-			{
-				return ItemViewType.Footer;
-			}
-
 			if (_usingItemTemplate)
 			{
 				return ItemViewType.TemplatedItem;
@@ -206,43 +155,6 @@ namespace Xamarin.Forms.Platform.Android
 		void UpdateUsingItemTemplate()
 		{
 			_usingItemTemplate = ItemsView.ItemTemplate != null;
-		}
-
-		void UpdateHasHeader()
-		{
-			ItemsSource.HasHeader = ItemsView.Header != null;
-		}
-
-		void UpdateHasFooter()
-		{
-			ItemsSource.HasFooter = ItemsView.Footer != null;
-		}
-
-		bool IsHeader(int position)
-		{
-			return ItemsSource.IsHeader(position);
-		}
-
-		bool IsFooter(int position)
-		{
-			return ItemsSource.IsFooter(position);
-		}
-
-		protected RecyclerView.ViewHolder CreateHeaderFooterViewHolder(object content, DataTemplate template, Context context)
-		{
-			if (template != null)
-			{
-				var footerContentView = new ItemContentView(context);
-				return new TemplatedItemViewHolder(footerContentView, template, isSelectionEnabled: false);
-			}
-
-			if (content is View formsView)
-			{
-				return SimpleViewHolder.FromFormsView(formsView, context);
-			}
-
-			// No template, Footer is not a Forms View, so just display Footer.ToString
-			return SimpleViewHolder.FromText(content?.ToString(), context, fill: false);
 		}
 	}
 }
