+++ Xamarin.Forms.Platform.iOS/CollectionView/ItemsViewController.cs
@@ -20,12 +20,6 @@ namespace Xamarin.Forms.Platform.iOS
 		UIView _emptyUIView;
 		VisualElement _emptyViewFormsElement;
 
-		UIView _headerUIView;
-		VisualElement _headerViewFormsElement;
-
-		UIView _footerUIView;
-		VisualElement _footerViewFormsElement;
-
 		protected UICollectionViewDelegator Delegator { get; set; }
 
 		public ItemsViewController(ItemsView itemsView, ItemsViewLayout layout) : base(layout)
@@ -68,21 +62,8 @@ namespace Xamarin.Forms.Platform.iOS
 
 			if (disposing)
 			{
-				if (_headerViewFormsElement != null)
-					_headerViewFormsElement.MeasureInvalidated -= OnFormsElementMeasureInvalidated;
-
-				if (_footerViewFormsElement != null)
-					_footerViewFormsElement.MeasureInvalidated -= OnFormsElementMeasureInvalidated;
-
 				ItemsSource?.Dispose();
 				_emptyUIView?.Dispose();
-				_headerUIView?.Dispose();
-				_footerUIView?.Dispose();
-
-				_headerUIView = null;
-				_headerViewFormsElement = null;
-				_footerUIView = null;
-				_footerViewFormsElement = null;
 				_emptyUIView = null;
 				_emptyViewFormsElement = null;
 			}
@@ -149,22 +130,6 @@ namespace Xamarin.Forms.Platform.iOS
 			{
 				ItemsViewLayout.ConstrainTo(CollectionView.Bounds.Size);
 				_initialConstraintsSet = true;
-			}
-
-			// This update is only relevant if you have a footer view because it's used to place the footer view
-			// based on the ContentSize so we just update the positions if the ContentSize has changed
-			if (_footerUIView != null)
-			{
-				if (IsHorizontal)
-				{
-					if (_footerUIView.Frame.X != ItemsViewLayout.CollectionViewContentSize.Width)
-						UpdateHeaderFooterPosition();
-				}
-				else
-				{
-					if (_footerUIView.Frame.Y != ItemsViewLayout.CollectionViewContentSize.Height)
-						UpdateHeaderFooterPosition();
-				}
 			}
 		}
 
@@ -276,55 +241,7 @@ namespace Xamarin.Forms.Platform.iOS
 			CollectionView.RegisterClassForCell(typeof(VerticalCell), VerticalCell.ReuseId);
 		}
 
-		bool IsHorizontal => (ItemsView?.ItemsLayout as ItemsLayout)?.Orientation == ItemsLayoutOrientation.Horizontal;
-
-		void UpdateHeaderFooterPosition()
-		{
-			if (IsHorizontal)
-			{
-				var currentInset = CollectionView.ContentInset;
-
-				nfloat headerWidth = _headerUIView?.Frame.Width ?? 0f;
-				nfloat footerWidth = _footerUIView?.Frame.Width ?? 0f;
-
-				if (_headerUIView != null && _headerUIView.Frame.X != headerWidth)
-					_headerUIView.Frame = new CoreGraphics.CGRect(-headerWidth, 0, headerWidth, CollectionView.Frame.Height);
-
-				if (_footerUIView != null && (_footerUIView.Frame.X != ItemsViewLayout.CollectionViewContentSize.Width))
-					_footerUIView.Frame = new CoreGraphics.CGRect(ItemsViewLayout.CollectionViewContentSize.Width, 0, footerWidth, CollectionView.Frame.Height);
-
-				if (CollectionView.ContentInset.Left != headerWidth || CollectionView.ContentInset.Right != footerWidth)
-				{
-					CollectionView.ContentInset = new UIEdgeInsets(0, headerWidth, 0, footerWidth);
-
-					// if the header grows it will scroll off the screen because if you change the content inset iOS adjusts the content offset so the list doesn't move
-					// this changes the offset of the list by however much the header size has changed
-					CollectionView.ContentOffset = new CoreGraphics.CGPoint(CollectionView.ContentOffset.X + (currentInset.Left - CollectionView.ContentInset.Left), CollectionView.ContentOffset.Y);
-				}
-			}
-			else
-			{
-				var currentInset = CollectionView.ContentInset;
-
-				nfloat headerHeight = _headerUIView?.Frame.Height ?? 0f;
-				nfloat footerHeight = _footerUIView?.Frame.Height ?? 0f;
-
-				if (_headerUIView != null && _headerUIView.Frame.Y != headerHeight)
-					_headerUIView.Frame = new CoreGraphics.CGRect(0, -headerHeight, CollectionView.Frame.Width, headerHeight);
-
-				if (_footerUIView != null && (_footerUIView.Frame.Y != ItemsViewLayout.CollectionViewContentSize.Height))
-					_footerUIView.Frame = new CoreGraphics.CGRect(0, ItemsViewLayout.CollectionViewContentSize.Height, CollectionView.Frame.Width, footerHeight);
-
-				if (CollectionView.ContentInset.Top != headerHeight || CollectionView.ContentInset.Bottom != footerHeight)
-				{
-					CollectionView.ContentInset = new UIEdgeInsets(headerHeight, 0, footerHeight, 0);
-
-					// if the header grows it will scroll off the screen because if you change the content inset iOS adjusts the content offset so the list doesn't move
-					// this changes the offset of the list by however much the header size has changed
-					CollectionView.ContentOffset = new CoreGraphics.CGPoint(CollectionView.ContentOffset.X, CollectionView.ContentOffset.Y + (currentInset.Top - CollectionView.ContentInset.Top));
-				}
-			}
-		}
+		protected bool IsHorizontal => (ItemsView?.ItemsLayout as ItemsLayout)?.Orientation == ItemsLayoutOrientation.Horizontal;
 
 		internal void UpdateEmptyView()
 		{
@@ -334,17 +251,7 @@ namespace Xamarin.Forms.Platform.iOS
 			UpdateEmptyViewVisibility(ItemsSource?.ItemCount == 0);
 		}
 
-		internal void UpdateFooterView()
-		{
-			UpdateSubview(ItemsView?.Footer, ItemsView?.FooterTemplate, ref _footerUIView, ref _footerViewFormsElement);
-		}
-
-		internal void UpdateHeaderView()
-		{
-			UpdateSubview(ItemsView?.Header, ItemsView?.HeaderTemplate, ref _headerUIView, ref _headerViewFormsElement);
-		}
-
-		internal void UpdateSubview(object view, DataTemplate viewTemplate, ref UIView uiView, ref VisualElement formsElement)
+		protected void UpdateSubview(object view, DataTemplate viewTemplate, ref UIView uiView, ref VisualElement formsElement)
 		{
 			if (uiView != null)
 				CollectionView.Subviews.Remove(uiView);
@@ -388,13 +295,17 @@ namespace Xamarin.Forms.Platform.iOS
 			}
 		}
 
-		void OnFormsElementMeasureInvalidated(object sender, EventArgs e)
+		protected void OnFormsElementMeasureInvalidated(object sender, EventArgs e)
 		{
 			if (sender is VisualElement formsElement)
 			{
-				RemeasureLayout(formsElement);
-				UpdateHeaderFooterPosition();
+				HandleFormsElementMeasureInvalidated(formsElement);
 			}
+		}
+
+		protected virtual void HandleFormsElementMeasureInvalidated(VisualElement formsElement)
+		{
+			RemeasureLayout(formsElement);
 		}
 
 		internal void UpdateView(object view, DataTemplate viewTemplate, ref UIView uiView, ref VisualElement formsElement)
